+++ functions/api/class/parseNotation.js
@@ -1,8 +1,13 @@
 var Peo = require('peo')
 
+var regex = require('../../constants/regex')
+
+
 var parseNotation = function(notation) {
   // Should output a Peo for notation
+
   return new Peo()
+  // Return value should be a Peo
 }
 
 module.exports = parseNotation
